+++ src/lib/db/migrations/0001_thin_baron_strucker.sql
@@ -1,0 +1,2 @@
+ALTER TABLE `accounts` ADD `tax_wrapper` text DEFAULT 'none' NOT NULL;--> statement-breakpoint
+ALTER TABLE `accounts` ADD `category` text NOT NULL;